Add src folder and separate logic
