+++ Dockerfile
@@ -5,7 +5,7 @@ FROM node:18.14.2
 EXPOSE 4000
 # Copies index.js, the two package files, and the spreadsheets from the local 
 # directory to a new app directory on the container
-COPY index.js package.json yarn.lock spreadsheets  app/
+COPY src index.js package.json yarn.lock spreadsheets  app/
 # Changes working directory to the new directory just created
 WORKDIR /app
 # Installs dependencies on container
