+++ src/constants.js
@@ -1,0 +1,38 @@
+export const enlistedUserColumns = {
+  Grade: 4,
+  DUTYTITLE: 8,
+  AMU: 9,
+  DOD_ID: 10,
+  ATP31: 11,
+  AFC291_01: 12,
+  AMF: 13,
+  MPF: 20,
+  MAJCOM: 22,
+  Country: 23,
+  BASE_LOC: 24,
+  Org_type: 25,
+  EOPDate: 30,
+  Last_Name: 32,
+  First_name: 33,
+  Middle_Name: 34,
+};
+
+export const officerUserColumns = {
+  ANB2: 6,
+  DOD_ID: 7,
+  ATP31: 8,
+  CAS3: 9,
+  AMF: 12,
+  Grade: 13,
+  DUTYTITLE: 16,
+  MPF: 21,
+  CMD: 23,
+  MAJCOM: 24,
+  Country: 25,
+  BASE_LOC: 26,
+  org_kind: 27,
+  EOPDate: 31,
+  Last_Name: 34,
+  First_name: 35,
+  Middle_Name: 36,
+};
